Refactor routes to routes & controllers (and extract mail into separate module - fixed picture embeding)
