+++ middleware/auth.js
@@ -1,5 +1,6 @@
 const dotenv = require("dotenv");
 const jwt = require("jsonwebtoken");
+const AppError = require("../modules/errorHandling/AppError");
 
 if (process.env.NODE_ENV === "production") {
   dotenv.config({ path: `.env.production` });
@@ -12,17 +13,18 @@ const verifyToken = (req, res, next) => {
   const authHeader = req.headers.authorization;
   const token = req.headers.authorization?.split(" ")[1];
   if (!authHeader || !token) {
-    return res.status(401).json({ message: "Unauthorized" });
+    throw new AppError("Unauthorized", 401);
   }
   jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, decoded) => {
     // add logic to detect if someone tampered with access token
     if (err) {
       if (err.name == "TokenExpiredError") {
-        return res.status(401).json({ message: "Access token expired" });
+        throw new AppError("Access token expired", 401);
       } else {
-        return res.status(401).json({
-          message: "Unauthorized, token signature usuccessfuly verified",
-        });
+        throw new AppError(
+          "Unauthorized, token signature usuccessfuly verified",
+          401
+        );
       }
     }
     req.user = decoded;
