+++ routes/bookings.js
@@ -1,49 +1,26 @@
 const router = require("express").Router();
-const Booking = require("../models/Booking");
-const Post = require("../models/Post");
-const User = require("../models/User");
+
 const { verifyToken } = require("../middleware/auth");
+const { tryCatch } = require("../modules/errorHandling/tryCatch");
+const BookingsController = require("../controllers/bookings");
+
+const bookingsController = new BookingsController();
 
 // Create a booking
-router.post("/", verifyToken, async (req, res) => {
-  try {
-    const { postId } = req.body;
-    const post = await Post.findById(postId);
-    const host = post.userId;
-    const booking = new Booking({
-      hostId: host,
-      ...req.body,
-    });
-    await booking.save();
-    res.status(201).json(booking);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: error.message });
-  }
-});
+router.post(
+  "/",
+  verifyToken,
+  tryCatch(async (req, res) => {
+    await bookingsController.createBooking(req, res);
+  })
+);
 
 // Get all taken dates for a specific post
-router.get("/post/:postId/dates", async (req, res) => {
-  try {
-    const { postId } = req.params;
-    const post = await Post.findById(postId);
-    if (!post) {
-      return res.status(404).json({ error: "Post not found" });
-    }
-    const bookings = await Booking.find({ postId });
-    let takenDates = [];
-    bookings.forEach((booking) => {
-      let date = booking.startDate;
-      while (date <= booking.endDate) {
-        takenDates.push(date.toISOString().split("T")[0]);
-        date.setDate(date.getDate() + 1);
-      }
-    });
-    res.json(takenDates);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: error.message });
-  }
-});
+router.get(
+  "/post/:postId/dates",
+  tryCatch(async (req, res) => {
+    await bookingsController.getPostTakenDates(req, res);
+  })
+);
 
 module.exports = router;
